Refactor dash handler into api
